kmg_autotrader/requirements.txt
python-binance
asyncpg
websockets
fastapi
uvicorn
python-dotenv
pydantic
openai
joblib
scikit-learn
apscheduler
